[Refactor] 커스텀 컴포넌트 타입 분리 (.types.ts) - 12개 커스텀 컴포넌트에 .types.ts 파일 생성 - 컴포넌트에서 인라인 타입 제거 후 타입 파일에서 import - index.ts에서 컴포넌트와 타입 명시적 re-export - i18n 지원을 위한 labels props 추가 (FileUpload, DataTable) 적용 컴포넌트: Carousel, Chart, Combobox, DataTable, DateRangePicker, EmptyState, FileUpload, Kanban, Pagination, StatsCard, Stepper, Timeline
