+++ src/components/common/Carousel/index.ts
@@ -1,1 +1,2 @@
-export * from './Carousel';
+export { Carousel, CarouselContent, CarouselItem, CarouselPrevious, CarouselNext } from './Carousel';
+export type { CarouselApi, CarouselOptions, CarouselPlugin, CarouselProps, CarouselContextProps } from './Carousel.types';
