+++ src/components/common/Chart/index.ts
@@ -1,1 +1,2 @@
-export * from './Chart';
+export { ChartContainer, ChartTooltip, ChartTooltipContent, ChartLegend, ChartLegendContent, ChartStyle, useChart } from './Chart';
+export type { ChartConfig, ChartContextProps, ChartContainerProps, ChartStyleProps, PayloadItem, ChartTooltipContentProps, LegendPayloadItem, ChartLegendContentProps } from './Chart.types';
