+++ src/components/common/Combobox/Combobox.types.ts
@@ -1,0 +1,28 @@
+export interface ComboboxOption {
+  value: string;
+  label: string;
+  disabled?: boolean;
+}
+
+export interface ComboboxProps {
+  options: ComboboxOption[];
+  value?: string;
+  onValueChange?: (value: string) => void;
+  placeholder?: string;
+  searchPlaceholder?: string;
+  emptyMessage?: string;
+  className?: string;
+  disabled?: boolean;
+}
+
+export interface MultiComboboxProps {
+  options: ComboboxOption[];
+  values?: string[];
+  onValuesChange?: (values: string[]) => void;
+  placeholder?: string;
+  searchPlaceholder?: string;
+  emptyMessage?: string;
+  className?: string;
+  disabled?: boolean;
+  maxDisplay?: number;
+}
