+++ src/components/common/Combobox/index.ts
@@ -1,1 +1,2 @@
-export * from './Combobox';
+export { Combobox, MultiCombobox } from './Combobox';
+export type { ComboboxOption, ComboboxProps, MultiComboboxProps } from './Combobox.types';
