+++ src/components/common/DataTable/index.ts
@@ -1,1 +1,2 @@
-export * from './DataTable';
+export { DataTable, DataTableColumnHeader } from './DataTable';
+export type { DataTableProps, DataTableLabels, DataTableColumnHeaderProps } from './DataTable.types';
