+++ src/components/common/DateRangePicker/index.ts
@@ -1,1 +1,2 @@
-export * from './DateRangePicker';
+export { DateRangePicker, DatePicker } from './DateRangePicker';
+export type { DateRangePickerProps, DatePickerProps } from './DateRangePicker.types';
