+++ src/components/common/EmptyState/index.ts
@@ -1,1 +1,2 @@
-export * from './EmptyState';
+export { EmptyState, NoResultsEmpty, NoDataEmpty, ErrorEmpty } from './EmptyState';
+export type { EmptyStateProps, NoResultsEmptyProps, NoDataEmptyProps, ErrorEmptyProps } from './EmptyState.types';
