+++ src/components/common/FileUpload/index.ts
@@ -1,1 +1,2 @@
-export * from './FileUpload';
+export { FileUpload, ImageUpload } from './FileUpload';
+export type { FileUploadProps, FileUploadLabels, ImageUploadProps, ImageUploadLabels } from './FileUpload.types';
