+++ src/components/common/Kanban/index.ts
@@ -1,1 +1,2 @@
-export * from './Kanban';
+export { KanbanBoard, KanbanColumnComponent, KanbanCard } from './Kanban';
+export type { KanbanItem, KanbanColumn, KanbanCardProps, KanbanColumnProps, KanbanBoardProps } from './Kanban.types';
