+++ src/components/common/Pagination/Pagination.types.ts
@@ -1,0 +1,6 @@
+import type { Button } from '@/components/common/Button';
+
+export type PaginationLinkProps = {
+  isActive?: boolean;
+} & Pick<React.ComponentProps<typeof Button>, 'size'> &
+  React.ComponentProps<'a'>;
