+++ src/components/common/Pagination/index.ts
@@ -1,1 +1,2 @@
-export * from './Pagination';
+export { Pagination, PaginationContent, PaginationLink, PaginationItem, PaginationPrevious, PaginationNext, PaginationEllipsis } from './Pagination';
+export type { PaginationLinkProps } from './Pagination.types';
