+++ src/components/common/StatsCard/StatsCard.types.ts
@@ -1,0 +1,27 @@
+import type { LucideIcon } from 'lucide-react';
+
+export interface StatsCardProps {
+  title: string;
+  value: string | number;
+  description?: string;
+  icon?: LucideIcon;
+  trend?: {
+    value: number;
+    label?: string;
+  };
+  className?: string;
+  valueClassName?: string;
+}
+
+export interface StatsGridProps {
+  children: React.ReactNode;
+  columns?: 2 | 3 | 4;
+  className?: string;
+}
+
+export interface MiniStatsProps {
+  label: string;
+  value: string | number;
+  trend?: number;
+  className?: string;
+}
