+++ src/components/common/StatsCard/index.ts
@@ -1,1 +1,2 @@
-export * from './StatsCard';
+export { StatsCard, StatsGrid, MiniStats } from './StatsCard';
+export type { StatsCardProps, StatsGridProps, MiniStatsProps } from './StatsCard.types';
