+++ src/components/common/Stepper/index.ts
@@ -1,1 +1,2 @@
-export * from './Stepper';
+export { Stepper, StepperWithContent, StepperNavigation } from './Stepper';
+export type { Step, StepperProps, StepperWithContentProps, StepperNavigationProps } from './Stepper.types';
