+++ src/components/common/Timeline/index.ts
@@ -1,1 +1,2 @@
-export * from './Timeline';
+export { Timeline, TimelineItem, HorizontalTimeline, HorizontalTimelineItem } from './Timeline';
+export type { TimelineProps, TimelineItemProps, HorizontalTimelineProps, HorizontalTimelineItemProps } from './Timeline.types';
