+++ src/components/common/index.ts
@@ -298,7 +298,8 @@ export { DataTable } from './DataTable';
 export { DateRangePicker } from './DateRangePicker';
 export { EmptyState } from './EmptyState';
 export { FileUpload } from './FileUpload';
-export { KanbanBoard, KanbanColumn, KanbanCard } from './Kanban';
+export { KanbanBoard, KanbanColumnComponent, KanbanCard } from './Kanban';
+export type { KanbanColumn } from './Kanban';
 export { StatsCard } from './StatsCard';
 export { Stepper } from './Stepper';
 export type { Step } from './Stepper';
